Notify VIS the detector callback remote exception occurred
It's possible the VIMS fails to notify client because of the IPC
exception error, some unknown reason that causes the binder object
may have problem to callback. We assume the system still have binders
that is available to use, we try to notify the client procrss not use
callback object.

If the system is in a bad state that cannot get any binder, we do not
have a good solution to resolve this case now. More details can be
found in go/b244391070. The RemoteException for VQS if not be
included in this change, we focus on the hotword part.

Bug: 244391070
Test: manual. Force to call the method to trigger call path. The
onError() callback is called.

Change-Id: I30186d7efdb14c9ef0c7ed65bc089b89199d2e16

Former-commit-id: 3eb525997c04bfe367ed09c23f5eb1e8dd2aac69