Remove Dream Overlay window name and update content descriptions.
This changelist sets the Dream Overlay window name to an empty string.
Omitting the name causes the window to be attributed to SystemUI by
Talkback for any visual element (status bar, complications) rather than
the dream.

This changelist also sets the content descriptions for a number of
icons in the dream overlay.

Test: manual
Fixes: 267837462
Change-Id: Ic31d2c53d53a3e4c961cfc4e5a01245822b22b96

Former-commit-id: 21b031eece01671ca066f317925d993ff987b996