Camera quick affordance respects device policy.
If the device admin policy disables the camera, neither the stills nor
the video camera quick affordances are available. If stills one also
respects the "secure camera" disablement flag, which the the video one
does not need to, because it always requires an unlock to access the
video camera (due to a missing camera API).

Bug: 268218507
Test: unit tests added
Test: temporarily faked DevicePolicyManager API and manually verified
that the camera and video camera options do not show up in wallpaper
picker.

Change-Id: Ia5b68283a25bc686ab11258a4b55b001fabcf562

Former-commit-id: afc6a41e88ff1e6528a9ef02d98690214eb012b7