Specify the parent viewgroup that should hide controls
Without this, launching ControlsActivtiy while it's running may cause to
two instances of it existing for a brief moment. As the new instance
will call `show` before the old instance calls `hide`, this will cause
everything to be hidden (and all listeners removed).

Instead, specify what parent view is calling hide. This makes it so we
can decide whether to completely hide `ControlsUiControllerImpl` or just
remove the views from the old parent.

Test: atest ControlsUiControllerImplTest
Test: manual, launch from LS and then from QS
Test: heapdump: old ControlsActivity is released
Bug: 267949140
Change-Id: I86a323cf25e5da81e610a21c4426872ef8684480

Former-commit-id: efa0a41ffec7b791de5bafe01b60d63afe7ce253