LockSettingsService: remove ensureProfileKeystoreUnlocked()
It used to be the case that on unencrypted devices and on Full Disk
Encryption (FDE) devices, when a user was started it would immediately
be unlocked too, even if it had a lockscreen credential.

http://ag/19599753 (commit c6130e5e6ee4 [formerly 3d5653e11ec8]) changed the flow to be the same
as on File Based Encryption (FBE) devices.  Now, users are not unlocked
until their lockscreen credential is verified.

This means that the workaround implemented by
LockSettingsService.ensureProfileKeystoreUnlocked(), which decrypted and
verified the profile password when a profile with a unified challenge
was being unlocked, is no longer necessary.  Remove it.

(Note: FDE is no longer supported anyway.  However, "no encryption" is
still supported, sort of, for device bring-up only.  This workaround
applied to both "no encryption" and FDE, so that's why the actual reason
it's no longer necessary is a bit more subtle than "this was FDE code".)

Bug: 123215607
Change-Id: I7a0b694ac9120e2e98e0a155fd53b5d590ed87af

Former-commit-id: c0676da2aa11a594c6dd11dccf6672ee97151a5c