Support handwriting delegation across different view trees
Handwriting delegation allows stylus movement on a delegator view to
initiate handwriting mode for a delegate editor view.

The first implementation of handwriting delegation did not work when the
delegator and delegate views are in different view trees. The
handwriting delegate APIs are updated here to support these use cases.

Additional APIs are provided to support the case when the delegator and
delegate views are in different packages. These will be used by the
Settings app.

Bug: 266834695
Bug: 228100684
Test: atest android.view.stylus.HandwritingInitiatorTest
Test: atest android.view.inputmethod.cts.StylusHandwritingTest
Change-Id: I878febe8841451597f534b464d2b0c794e86bba8

Former-commit-id: 2680cc740a2cb0c2fb3c4464a3f906d32b1168f4