Check config change for reporting resize and requesting redraw
Since multi-window mode is supported, the orientation change is
not the only source that requires redraw. It is confusing to use
"set orientation change" to handle other cases that also need to
redraw. So simplify the condition to check if there is a config
change of the window. While drag-resizing is similar, the end of
drag-resizing should also trigger a config change.

Also
- Reset mLastConfigReportedToClient only if the config is changed.
  This can avoid unnecessary redraw request if a window is reparented
  to another container with the same configuration.
- Set mWaitingForConfig for 180 rotation because since
  WindowConfiguration was introduced, there will be always a
  configuration change for rotation change.
- Add trace for global config change.

Bug: 159103089
Test: WindowStateTests

Change-Id: I50d4c5b741be960bd2850e4bcd0ab62ad32c19dc

Former-commit-id: 31a3779b15c47aea060c7bf37bb09d05437538af