Allow to move SCM Letterbox with Taskbar visible
Currently SCM reduces height of the letterbox by the
height of Taskbar, due to this letterboxuicontroller
believes that the application does not take the full
height of the screen.

This CL fixes this issue by using getAppBounds()
instead getBounds() to query the height of the available
space on the screen.

Change-Id: I152b588bc5f3673e0e2aa6e0a43d5e40ad234560
Test: atest SizeCompatTests , also manual: SCM letterbox can be moved horizontally
Bug: 269030028

Former-commit-id: d5f15e2e354df2a900cdf46ee559983138f0d0f7