Rename summaryText where it's not actually related.
summaryText is a property of BigTextStyle notifications, so it's a bit
confusing to reuse the same name for things that are... not that.

Also had to add an m at the beginning of this and related properties
since checkStyle was complaining, and it's now consistent with other
properties.

Test: presubmit
Bug: 269474373
Change-Id: I75d5463fd67afddb4e8f0acc9c12e0a2c8819013
(cherry picked from commit 68dda71f7a1274111801792f5163235978c91832 [formerly 2311260ab38b098af9746c85930f02ed5599cba2])

Former-commit-id: 5a8f7baeeb62ca44114c74987e531c3efef09cd4