Add a fallback check to prevent half-way pip anims
This situation leaves the device in an unusable state, so
add a hard-check for this to recover if it happens unexpectedly.
This code should never run; however, just to be safe, its
preferable to fallback and log an error instead of make the device
unusable if an edge-case does get hit.

Bug: 267897440
Test: purposefully disable some pip handling in shell and verify
      that the fallback occurs
Change-Id: I619b49d5c4dedd92fac5ea3b8e133c8c50a72d9b

Former-commit-id: 66a2adc46af2b2a9d92fe0161ec083eb9a5d159c