Enable protolog by default for WMShell
WMShell is a static library built into systemui. In this change, we
allocate a 128KB buffer and enable it by default on debuggable builds,
e.g.  userdebug or eng builds. This should give us sufficient amount of
log for debugging purpose when receive bug-reports.

We need further have the viewer in bug-report / winscope to actually
view the logs being written.

Bug: 262244460
Test: adb shell dumpsys activity service SystemUIService WMShell \
      protolog save-for-bugreport
Change-Id: I50e1ad31d944ac530c9c7f2d881c09cf05493be9

Former-commit-id: 4005c0e25c49213834b8f94ce1ba98a35e40246f