Made TaskViewTransitions less rigid
Because, at the current time, its possible for non-taskview
changes to end up in a taskview transition, we have to make
taskview transition-handler capable of doing housekeeping on
just the taskviews and ignoring the other stuff.

This requires using the launch-cookie so that it can differentiate
taskview launches from unrelated ones.

Also added an onConsumed impl so that pending taskview transitions
don't get "leaked" if they end up as no-op (aborted).

Bug: 268907067
Test: existing taskview tests pass. Open the Home settings from
      quick-settings, then a device, quickly exit both.
Change-Id: I856a2d2fbc1f338f5a2e1d0a14a61a646d6ecbbf

Former-commit-id: 692859e534b4eea20e8ed4a2b568667d4a094b52