Hard code caption insets for caption in WM shell
We attempted to remove the hard coded caption insets for captions in WM
shell before, but it revealed a synchronization issue of window frames
and caption inset frames being updated separately. The proper fix to
this issue is more complicated than I initially thought, so we restore
the workaround in a similar way as the workaround we had for the legacy
DecorCaptionView in short term.

For now InsetsPolicy drops caption insets if the the layout attributes
indicate the window is floating.

Also restore unit tests removed because the old workaround was removed
for caption in WM shell.

Bug: 254128050
Test: Caption insets look correct for regular activities.
Test: Floating windows (dialogs) don't get caption insets.
Test: atest InsetsSourceTest
Test: atest InsetsStateTest
Change-Id: I15616974bbccc8662124e9efa956c180dc329cbf

Former-commit-id: 849d65f38c4272a05bd78afbd0d4d156357fc818