feat(non linear font scaling): interpolate between two tables if one cannot be found for a given scale
If a new scale is defined in the OEM’s Settings XML that doesn't match
one of the hardcoded default tables, one will be generated by
interpolating between two default tables. This way the experience is
guaranteed for the user, even if the OEM forgets to define their tables
to match their font scaling settings.

Test: atest FrameworksCoreTests:android.content.res.FontScaleConverterFactoryTest
Bug: b/247861716

Change-Id: If9c7f87f6dd4b99afa8579cfa0b52e313d761beb

Former-commit-id: de466db1a5b20095532622487363a0632648e335