Use a weak reference to the context in InputManager
It has been reported that a recent change to InputManager is
causing a memory leak. We were already planning to change it in a
way that this issue would no longer happen, but since it's a
refactor that will take some time, this is a temporary solution
that Prabir suggested; if we keep a weak reference to the context,
it will be freed when there is no other active references.

Bug: 267241786
Test: will ask for a test run of this change
Change-Id: I7e894ba1f014bba36f6ea2566b11426d17d84f68
(cherry picked from commit 26608414af0922e567111d76d95df27cf95cd0e2 [formerly d87354712e460cad920a8f09df3133fa257cf283])
Merged-In: I7e894ba1f014bba36f6ea2566b11426d17d84f68

Former-commit-id: 71feb32d93cefbf6921034001b856742dde9ff9c