Fixes failing CustomizationProviderTest.
The issue was the recent introduction of runBlocking { } calls in
ag/21396218. These block the logic on the current thread, causing unit
test runners to choke. Instead, we should use
runBlocking(CoroutineDispatcher) { } to make sure that the suspend
function yields back to the running test.

In addition to that change, we also needed to move the test onto an
UnconfinedTestDispatcher.

Fix: 269679515
Test: tests pass locally with atest
Change-Id: Iff661303af8791df7751fac7788681a105d47a75

Former-commit-id: 3528f2121655873b75142d698d63a77137ba7ff7