Do not use sync on divider visibility changed.
When we set divider visibility, we only use sync transcation on
hidden case because we need do fade-in animation for show case.
This might lead divider hidden on some race if  we set hidden at
first and show later case because the first hidden is waiting sync
and apply later actually.

Fix: 269049102
Test: manual
Test: pass existing tests
Change-Id: Ibdb91cc5c8e1d3a1f9c3726df7d467c7255998f0

Former-commit-id: b0261472618c1c2c85d748ba689dff1781fcc555