Fix unlocked split shade jumping back into place after over expanding
The problem was that there were separate code paths for over-scroll/
over-expansion for split shade and non split shade.
In this case the non split shade path was altering translation at the
end of the drag, even when on split shade.

To avoid adding even more "if split shade" code, removed the split shade
over-scrolling code path/classes.

There is only one overscroll code path now, with just 1 "if split" shade
condition to decide which components to over expand.

Test: Manually on device. See videos.
Fixes: 270005948
Change-Id: Ia47aa30957260a9c2d1efd7e5f18536f76a2dcc9

Former-commit-id: cafd48c62c74e40d29c6a90e224aa3eb2fe7ab95