Fix a11y announcements in DND duration dialogs
We are using a separate RadioButton just for the button and a view (possibly complex) fo the text. The latter didn't report an accurate selection status (especially when selecting, which didn't announce anything at all). Note that we are _not_ setting the "unselected" status to avoid repetition when walking through with Talkback (button -> text -> button -> text, ...).

Fixes: 265468127
Test: Unit tests
Change-Id: I87aad93d942a868776ecec163136fd9bfcb8276f

Former-commit-id: 3d3ef50243a1ceed2c03653bae271b9eb23da804