Implement IMMS#mCurFocusedWindowEditorInfo
This fixes an issue where the first IME request is a hide, and thus
no mCurEditorInfo was set, but a non-null editorInfo was present
in IMMS#startInputOrWindowGainedFocus. This is used to get the
packageName for ImeTracker.Token; as mCurEditorInfo was
stale (for hide requests), it would lead to misleading logs.

Bug: 261565259
Test: filter logcat by "ImeTracker"; observe first request after SysUI
  boots (trigger by adb shell stop; adb shell start). This should now
  reflect packageName of settings (was null before).
Change-Id: I296884140cbcb8511e9e9a5d03a0ff4092859433

Former-commit-id: 3f139b669e471ad4182cd729a9ca420356465726