Handle invalid dream window in DreamOverlayService.
It is possible for a dream's window to become invalid shortly after
starting. For example, another dream might begin almost immediately
following. In these cases, it is possible that the DreamOverlayService
might not process the start event until after the dream window becomes
invalid. If not handled properly, an BadTokenException will cause
SystemUI to crash.

This changelist addresses the issue by catching the exception and
tearing down the overlay.

Test: atest DreamOverlayServiceTest
Fixes: 268209056
Change-Id: I25ff6ff58e6a612ea609a4541219bea149eeace9

Former-commit-id: b0514b19d841d64f22274cac7948f81c1bfa68fc