Make RemoteEntry flow independent
Based on API feedback from DP1 testers, we need to make RemoteEntry a standalone feature
irrespective of the type of the credential, or a particular option from the GetCredentialRequest.
With this change, remote entry can be created with just some UI metadata (defined on jetpack side).
When user selects this entry, we send all CredentialOptions to this provider and the entire list can be
serviced on a remote device.
Consequently, this change includes modifying android.service.GetCredentialRequest to take in a list of
CredentialOption's as opposed to a single CredentialOption. When a CredentialEntry is selected, that list
will always contain a single option. When a RemoteEntry is selected, that list may conttain >1 options.
A RemoteEntry is  not tied to a type and hence we need a dedicated type for it that only contains Slice (holding the UI metadata). Earlier we were using a CredentialEntry to represent a remote entry in the get flow. However a CredentialEntry requires a type and hence is not applicable. In the create flow, we were using a CreateEntry which does not require a type but in order to keep the flows symmetric, we decide to use RemoteEntry as a separate entry type in both flows.

Test: built locally & CTS tests
Bug: 270186121

Change-Id: Ibfe76d88bde727cb8eed3227ae0dc26ae7fe6834

Former-commit-id: 48c0a642e7d228cd765b849cfd4d7681cac0d642