Restore empty ZenRule#conditionId after a device restart
We allow to create a new ZenRule with an empty (not null) conditionId. When restarting the device, we restore these rules form an XML. We didn't restore an empty conditionId. It lead to skip restoring the whole ZenConfig containing a rule with empty conditionId, which resulted in deleting all the rules in the given config.

Fixes: 264663537
Test: ZenModeConfigTest
Test: create some ZenRules manually form the Settings app, create a new ZenRule with an empty conditionId and add it with NotificationManager#addAutomaticZenRule using a test app, restart the device, see that all the rules are restored
Change-Id: I2e00c79216734f6c2db7d5b1347ab72a9be5cdc3

Former-commit-id: 1cc88e774235feb4d9f89c9423365de10018e47d