Always execute logic to move desktop apps to front
We had an optimization in bringDesktopAppsToFront to skip it when we
detect that all tasks are already visible.
This method was being called in two cases, when showing desktop tasks or
when moving an existing task to desktop.
We were already skipping this optimization when showing desktop tasks
using the force arg.
Also, when moving a task to desktop, this optimization was not actually
used. As we only move a task to desktop when switching a fullscreen task
to desktop. And in this case no other desktop task is visible.

Removing this optimization also fixes a bug where moving a task to
desktop can result in a fullscreen task remaining visible in the back.
The reason is that bringDesktopAppsToFront also is responsible for
ensuring that the launcher task is at the right location. And the
optimization caused us to skip it.

Bug: 269663000
Test: atest DesktopTasksControllerTest
Change-Id: I8d9ba762c22ea1fed5fcf955b7098d8747cd553a

Former-commit-id: 833db3b471d0594cbbd38f47e8ae88f821c6ea53