[People Service] Add throttling to DataManager add / change listener
We add throttling to the `onShortcutsAddedOrUpdated` shortcut service
callback within the `DataManager`, as well as re-order the clean-up
function to only execute once per update in order reduce the load on
People Service and system_server. This is especially important in
cases where apps publish many conversation shortcuts in a short period.

Change-Id: I04b0c42dabfa4f1885bea2cf6efb3c4693220585
Test: unit tests and manually via load test apk
Bug: 269420266

Former-commit-id: d5b37667d0b36c9fc55f27b2783c295826d59026