Fix wrong ime parent in embedded activity(2nd try)
We have CL[1] to polish the ime transition. But in
ActivityEmbedding, the ime layering target may higher
than input target. And the ime parent didn't be updated
since input target and layering target are unsync.

Update ime parent for this case.

Also we ensure both ime targets are attached and are in
same embedded task before window hierarchy comparison to
prevent the exception.

[1]: I332c0e4fff62df5d7b793eda2767bb58fe85a938

Bug: 260387203
Bug: 270290011
Test: Manual test with test apk in the bug
Test: Verify no exception through bug steps
Test: atest TaskFragmentTest#testUpdateImeParentForActivityEmbedding
Change-Id: I937560f7479535f0e2dc77ebe0f5056fa85bbb04

Former-commit-id: fc837a1811443cc9eb4a18e5c6627a183df2b352