Disable back gesture on the home and -1 screen
The way to disable back on home / -1 screen is by specifying the entire
screen as part of systemGestureExclusion when launcher is in that state.
For trackpad gestures, we bypass that check generally when the cursor is
in an excluded region (since 3-finger swipe doesn't interact with that
region instead of touch). However, in a case like this (when the whole
screen is excluded from system gesture), we should respect it.

Fixes: 270585141
Test: 3-finger horizontal swipe from home / -1, make sure that back
behavior is not triggered

Change-Id: I329fc0f3efbb268e2fa1a9e9e33902cd523d4253

Former-commit-id: e8008e814aba547f9aa51a274a2fe4943c71860e