Play seekable animation for customize activity transition API.(2/N)
Add supporting for Activity#overrideActivityTransition.
The CustomizeActivityAnimation will load both enter and exit animation
for the close activity transition. And it is only valid if the exit
animation has set and loaded success. If the entering animation has not
set(i.e. 0), there will load the default entering animation for it.
Also note that if both overrideActivityTransition and windowAnimations
has set, system should prior to load the animation set from
overrideActivityTransition.

Bug: 259427810
Test: atest ActivityTransitionTests BackNavigationControllerTests\
CustomizeActivityAnimationTest

Change-Id: Id182dc8f93f55a256de68c1fb6cc91fd08450e3e

Former-commit-id: 850d69d3c8ef759e17d1c69fe602bb53ac92055a