Limit resizing to within the stable bounds
Resizing into the navbar area caused the task's bottom resize
handles to be unaccesible for new drag-resize interactions because
the touch/drag would be stolen by the system gesture detector to
swipe up.
This change disallows the task to be resized into the gesture area
in the first place, so that there's no overlap between the task's
edges and the gesture region.

Note that it is still possible to run into this overlap if the task
is moved, not resized, into the gesture area. Whether the task
should be forbidden from moving there (and offscreen), or if the
gesture should add the task edges as 'exclusion' regions is TBD.

Bug: 269659818
Test: atest WMShellUnitTests
Change-Id: I2dfe037062dad46db93e0e54c94e0cfdafa76bf6

Former-commit-id: 57857ab61c92ebeca39b2181dae3d91c59e13411