Accept a Window as shown if it's in any transition
Previously, a malicious or confused app can DoS all other apps with
unshown surfaces by kicking off at least one new transition every
frame.

For example (see linked bug), an app in a busy-loop trying to request
permissions can prevent the launcher from showing Recents.

The significant change here is that when we see that there is a
transition set, we only block on mToken.waitingToShow if mToken
is not already part of another transition. Otherwise, there can
be a new transition generated every frame and the token is constantly
stuck waiting to show.

Test: atest CtsWindowManagerDeviceTestCases
Bug: 269108443
Fix: 271085429
Change-Id: I47ce9fdf5429795cbf083516c19b669bc407d1da

Former-commit-id: d98e95c7c8892e45ca018b7c7227e3b658763df1