Always resume view when relaunching with preserved window
This relaxes the condition added in [1]. Because
IWindow#dispatchAppVisibility(true) may arrive after the
relaunch message, then getHostVisibility() will be GONE
because mAppVisible is still false.

The original concern is not a problem: call Activity#recreate()
while the activity is invisible. Because a stopped activity
won't use preserveWindow=true, it won't call VRI#notifyChildRebuilt.

[1]: Iee780c12535217a5a9726f9f4d6b21952ce56d70

Bug: 269179932
Test: WMShellFlickerTests:CopyContentInSplit

Change-Id: I7ea7d54f0c6bbf67b295ac3a2c97a62d81b0458b

Former-commit-id: ab34d09bb8ae4f34fda50e4222de74c8786daa8f