Ensure keyboard is hidden when bouncer is...
not showing. There are still cases where the keyboard will show. The
most common case that I have observed is that when the falsing manager
closes the bouncer after detecting a potential false touch. This can be
done by clicking on a notification, and double tapping the alternate
boucner overlay twice very quickly. The issue is that when we send a
signal to show the keyboard, there is no way to stop that signal before
the keyboard shows (from my understanding).

The solution here is to listen to onApplyWindowInsets, which is called
when the keyboard shows, and hide the keyboard if the bouncer is not
showing.

Fixes: 266816405
Test: Test as described above.
Change-Id: I71419781ab91d8a1404d3571cd7e7d16b2a3afc9

Former-commit-id: 8d0f687b31bb8eb35955e7904794ff2dd5a22483