Provide config for customize the showing strategy of back target.
Provide config config_predictShowStartingSurface with default = true.
If config_predictShowStartingSurface is true, system will always create
a snapshot or splash screen surface for the predice animation target.
When set config_predictShowStartingSurface to false, system will always
make the back animation target alive.

OEMs can choose the preview strategy for the opening target by override
the config based on the device, but in general, it should only be
disabled on low-ram device.
There will be a CTS to verify this behavior.

Bug: 268563842
Test: atest BackNavigationControllerTests
Change-Id: Iaa3a01b9852756d7096bba9a356f723b0c53f7cd

Former-commit-id: 1b7ef7b04a30d7ab2633806a8c82e86e1ea5ea2f