Make CEC HAL support required for eARC support
Reason is that the eARC implementation requires methods from the HDMI
Connection HAL, which currently enters the framework through
HdmiCecController. We may split this out in the future, but with the
current implementation, eARC cannot be supported if CEC is not
supported. Not that eARC can still work if the CEC setting is
*disabled*.

Test: make and atest
Bug: 240388106
Change-Id: Ie025c60c27e9bcc227d082e19782c674c7d7ad8b

Former-commit-id: 8b14b52d3e53b752b47d76290cb9ce8f9a615282