Ignore the closing MotionEvent of swipe gestures for falsing
The ZigZagClassifier falsly reports some swipes as false touches. The reason for that is in our touch event stream. The closing ACTION_UP event of a swipe might be a bit offseted from the previous ACTION_MOVE event and sometimes it is just enough to  push us over the allowed tolerance.

In this fix we drop the gesture closing ACTION_UP event if it is close in time to a previous ACTION_MOVE event.

Fixes: 236197248
Test: atest swipe notifications on the LS and observe the FalsingManager logs
Test: atest SystemUI:com.android.systemui.classifier

Change-Id: I33093ff77fd23a2913bd3adfd79a15e3ae66b316

Former-commit-id: 35a81def08e9cb2810b9f2bc96bb04f5210b7006