ASM - Allow sources to launch into own task with NEW_TASK
This was previously blocked to try and provide consistency with launching
into a different task. Relaxing for compatibility.

Usecase: An app launches its main_activity, which trampolines
with NEW_TASK and matching affinity to main_activity. If the user
goes home before the trampoline, it would be blocked before this change.

Bug: 270680647
Test: atest ActivitySecurityModelTest
Change-Id: Ib9412d6cd9c7183da19fee8afd4bc47524d5736c

Former-commit-id: 5cc88e94fd980cdbb4f7693ec4d5af78f642879b