locksettings: clean up logging of password history updates
Currently the logging for password history updates consists only of the
message "Initialized lock password salt for user".  But that's only
logged on the first update, and it's unclear that it's related to the
password history.  Let's replace it with a clearer message that is
logged whenever a password is added to the password history.

This doesn't change anything for the case where password history is
disabled, which is the default setting.

Bug: 268526331
Change-Id: Ibe6f679a17b887261711e0fd7da1b62a114ce7e6
Merged-In: Ibe6f679a17b887261711e0fd7da1b62a114ce7e6
(cherry picked from commit 61729d86bab407da02d8eeaac1b8f5893f0ae83b)

Former-commit-id: 0889b6d805f6c3ac85ddf000d604bbfb864193d3