MainUser permanent admin by default
On HSUM builds, we generally want the MainUser to be a permanent admin
user in AOSP. OEMs are free to turn that off, but then they are
responsible for handling initial user creation themselves.

This doesn't have any effect on non-HSUM builds (unless they are using
adb to emulate HSUM builds).

Bug: 270961661
Test: atest UserManagerTest
Change-Id: Iecf7129a9faeeede1bc1b8b7083dc98f37569556

Former-commit-id: f62d2aab119ced7364692fd0bedd99b5fb7de1e1