Require CREATE_VIRTUAL_DEVICE permission to register or unregister an AccessibilityDisplayProxy
This permission is only available to CDM apps with certain internal roles.
When associating with a companion device, a dialog asks users to approve
access, and on success this grants the CREATE_VIRTUAL_DEVICE permission.
To prevent abuse of accessibility APIs, accessibility will leverage this
existing flow for callers who want to register or unregister a proxy.

Bug: 265858731
Test: atest AccessibilityDisplayProxyTest, manually test the association flow
Change-Id: Ia2091b4a0f358cf4fe4aa624f69ec633523e1b36

Former-commit-id: 9e78be15cf19de3ab8697a7945aa3e96d0d36b2f