Improve event time crash logging
We got a report that system server is crashing due to receiving
some events with a time set in the future. From the error message
it is not possible to determine if the event time is a bit further
or if it's just uninitialized garbage (the second is a lot more
plausible). This CL will add some extra logging when this occurs
so we have a better idea of what is happening.

Bug: 269510424
Test: builds, passes presubmit
Change-Id: I47a924f182444dcdefea5b2cda76f7d2d8de1cf0

Former-commit-id: b56a64a37ed1ef10b932a489927ac2ad3d3d71e1