Consider entity length when deciding whether to show actions
Currently, we show the highest-confidence action, regardless of how
much space it takes up in the total copied text. This leads to some
strange behaviors like showing a calendar action if a date is shown
anywhere in the text.

This change adds a minimum proportion (currently 80%); if the text
link takes up less than that proportion of the total copied text,
it is ignored.

Bug: 270445121
Bug: 270421442
Fix: 270445121
Test: manual (using an article headline+date, as in the bug);
atest ClipboardOverlayUtilsTest

Change-Id: I99b9c48aae121244e01f8783d9eb8068d3c12fe3

Former-commit-id: 4ebe03cd0f7590ae333b60cee486d7ce8a1568e8