Change Size Compat mode resizability heuristic to look at top activity
Size Compat mode should be based only on top activity properties to improve debugging and predictability for the developer. The current policy of looking at the root activity was too aggressive and counter intuitive to developers, so we are changing the heuristic for resizability to look at the top activity instead. Now it will be easier to predict through the code analysis in an IDE that an activity will go in SCM since that will depend only on activity properties defined in the manifest and not on the order in which activities are launched.

Fix: 227795096
Test: manually with a test app
Test: atest WmTests:SizeCompatTests#testShouldCreateCompatDisplayInsetsWhenUnresizeableAndSupportsSizeChangesFalse

Change-Id: I5f57936ce65ceb5052d78398f3602eae669ea3cb

Former-commit-id: abc62403870f4d47ac660c784b3b4d3bd31d00d5