[People Service] Re-order clean-up to only occur once per shortcuts update
This is a partial roll-forward of the functionality added in ag/21546169.
We remove the 500ms debounce as this conflicts with direct boot aware
services, and we see an uptick of crashes caused by querying shortcut
manager when the user is locked. However, we keep the efficiency of only
cleaning up shortcuts once per update -- prior to this change we had
performed cleanup multiple times per update which created multiple heaps
in a single update operation.

Test: unit tests and manually via load test apk
Bug: 269420266
Change-Id: I9a445dd88e694c36f0c66491255481df9e4856f3

Former-commit-id: 68b75d0ceffd98599aee76c310d9298991d8b521