Replace View elevation with SC#setShadowRadius in handle menu
Using elevation on root views causes an input x-offset where clicks
and touches are not correctly aligned with the view's actual position,
which ultimately results in buttons being hard to click on the right
side.
SurfaceControl#setShadowRadius has the same effects as elevation
without the need to calculation the offset and without this bug, so
use that instead.

Bug: 270412904
Test: manual - use a mouse to hover over windowing buttons, verify
that ripple effect and clicks work over the correct bounds of the
button views

Change-Id: I325860676c88439fb00ed177d9897088eb09c730

Former-commit-id: ec3a11ac233de9e1cfa2c519629ebaed8754e211