Update launch policy state when finishing transient launch
This aligns with the legacy implementation of
RecentsAnimation#finishAnimation:
If it is returning to home, it will call stopAppSwitches
and moveTaskToFront -> TDA#onChildPositionChanged
-> RWC#invalidateTaskLayers.

Legacy: move home to top at the end.
Shell: move home to top at the at the beginning.
So manual call rankTaskLayers for shell transition.

Then the background launch restriction can detect the valid states
for the launches after the swipe-to-home gesture is done.

Bug: 270383113
Test: atest TransitionTests#testTransientLaunch

Change-Id: I0987f2a0afa2bf599322d37acfe19d40060f7879

Former-commit-id: 93845662e72104484455bbca1dad74829dbd2dcd