Send ACTION_SHUTDOWN as unordered broadcast with callback.
This broadcast was being sent as ordered in order to get the
finish callback. Since the modern broadcast queue supports
sending the finish callback without the need for having the
broadcast marked as ordered, convert this broadcast to be
sent as unordered with callback.

Bug: 253226131
Test: TH
Change-Id: I744884d846a9bdffca424ddbc07f9016d9300d02
Merged-In: I744884d846a9bdffca424ddbc07f9016d9300d02

Former-commit-id: 7a42cc181acf82b7949762585299146877657f78