Update OoB TalkBack tutorial with appropriate description of holding down volume buttons
Root cause: When a user goes to set up a device for the first time, they trigger the volume key shortcut to turn TalkBack on. After about three seconds of holding down the volume keys, they are prompted to continue holding down the volume keys to turn on accessibility. This prompt is incorrect, as users must actually release the keys and press again for three seconds to turn TalkBack on.
Solution: The tutorial should change to alert users that they must release, then press the volume keys again for three seconds, else the behavior of the volume key hold-down should change to mirror the tutorial.

Bug: 181575263
Test: Manual testing
Change-Id: Icc66c9a2cd45fb47bf658a319cc7745a36958d6e

Former-commit-id: efaa810d14eba337b7a87cae701930f539d790bd