Further V2 Modifications, Prep for V3
This adds to our Metric Objects the ability to handle new V2 changes,
and prepares for V3. There will be significant changes to come, but this
is a first step that addresses our immediate TODOs, particularily to
properly split up the latencies of the Metric objects on top of the
previously checked in isEnabled additions and the cancellation changes.

Bug: 270403549
Bug: 269290341
Test: Will be chained in the future
Change-Id: Ia77ef7fe854a2b41ef2bc85a07f3852df1984bbf

Former-commit-id: b349b59f7a8c82ff6dca25efa8146c3094e5e48e