Fix potential deadlock in transactions
It was possible to hit a deadlock scenario when using the
waitAsyncStart utility. Since that was using the main handler
for unlocked work, if the transaction was applied from the
main handler thread, it would deadlock (this happens when
transactions are queued-up on syncengine). This adds a check
and will run the uncommon queued launches on anim-thread
(which is 1 higher priority).

This also has the waitAsyncStart function actually return
the result so that we can properly observe the effects of
a PendingIntent

Bug: 271317661
Bug: 267738124
Test: System still runs and tests pass.
Change-Id: If99668b71cafe016b1ab2aad501c53d246e2877d

Former-commit-id: c8ecacc046492399b13719578c8f39b1b67e08b7