Use non-cached value when calculation button forced visible state
The change is to make use of the button forced visible state in
EdgeBackGestureHandler instead of the cached value in the Navigation
bar. The value was only available through a callback and the cached
value may be unpredictable when the value has not changed since system
boots. That may cause the insets provided by the navigation bar wrong.

When the device firstly boot into setup wizard the update user resource
and the callback is not called.

Keep the callback to reposition the navigation bar when necessary, but
only use read value from the EdgeBackGestureHandler to fix the issue.

Bug: 268210882
Test: Check setup wizard as stated in b/268210882
Change-Id: I31c6e06c0b1ab07aa8c2fdcacce216ca3ad61869

Former-commit-id: 4ad1250bf5830f914bd171955f23de5a7a7d38e4