Use a SLEEP transition to trigger failsafe animation cancelling
Some remote animations can misbehave. In particular, recents. When
this happens in the form of the remote not reporting finish, the
transition system can get stuck waiting. Legacy recents had the
same problem and solved it by cancelling the recents and
independently finishing/cleaning-up after a "failsafe" duration
from turning the screen off.

This CL sets-up something similar for shell-transitions. It uses
SLEEP as a signal to "quickly end all animations". This signal
gets sent (via merge) to all the animations. Each animation gets
a small amount of time to finish after-which we forcibly finish
it.

Bug: 267738124
Test: Observe sleep transitions and cancelling in logs.
Change-Id: I623fb5a47a5593e23fbc8f2298eff1c44ae3a0fc

Former-commit-id: 6831703c4aa115b1f772a2156bef1cea0ed0a6fa