Modify current KCM files to support shift+capslock
At the moment users will not be able to use lowercase keys
while caps lock is active. This was inconsistent, as the generic
layout supported it, while the rest did not.

Both Windows and ChromeOS allow using lowercase letters while
caps lock is on by pressing Shift. This change adds the necessary
mappings to implement the same behavior in Android.

While going through the files, I cleaned up some inconsistencies
and unnecessary rules.

Bug: 271341093
Test: flashed, tried several layouts
Change-Id: Idbafbf45e6db52838728021efc8e643b589c08e4

Former-commit-id: 9d3af2b9d66730ad6f4aaed53a497b1772dacdda