[Reland] VCN: Require capability matching on test networks for CTS
This commit updates NetworkPriorityClassifier to also require
capabilty matching on test networks. This allows CTS to select
underlying test networks based on their capabilities. Previously
capability matching was only done cell networks.

To do that this commit adds an abstract method to
VcnUnderlyingNetworkTemplate to retrieve capability requirements
so that the capability match can be done before the matching logic
diverge to different transports.

Bug: 263415068
Test: atest VcnManagerTest (new tests)
Change-Id: I12377cbd1d721c4100a958fd9a0ea56d87351eec

Former-commit-id: 191162949d6c551a206091a27ac5c48685a3a6eb