VCN: Require capability matching on test networks for CTS
This commit updates NetworkPriorityClassifier to also require
capabilty matching on test networks. This allows CTS to select
underlying test networks based on their capabilities. Previously
capability matching was only done cell networks.

To do that this commit adds an abstract method to
VcnUnderlyingNetworkTemplate to retrieve capability requirements
so that the capability match can be done before the matching logic
diverge to different transports.

Bug: 263415068
Test: atest VcnManagerTest (new tests)
Change-Id: Id6ee75e70dc31851d7d6980ee1bf5eceb843b294

Former-commit-id: 4477537461c864933093d21f9f716ab9bc5b0df0