Prevent ST clients from auto-attach to fake HAL
The current behavior is to pick the first ST module.
Right now, on devices with no HAL, clients use the fact
that the module fails to attach to detect lack of DSP presence.
We must retain this behavior when we add a second ST module
for testing, which is available on all devices.
Only select HALs not implemented by android for default construction.

Test: NowPlaying/HeyG attach verification
Test: When fake STHAL is added, NowPlaying/HeyG do NOT attach to it
Test: When real STHAL disabled, same as above
Test: getDspModuleProperties/getModuleProperties returns null on no real HAL
Bug: 271345481
Change-Id: Ie739eef56c4dedc2f7a79ec49f35793c95ca1658

Former-commit-id: 42b0291dabf6ec7ba857f8d9a339e106bb4d2487