fix(#Magnification): check activated instead of scale to notify magnifier when transitions
In AccessibilityController, when transitions we adjust to check isForceShowingMagnifiableBounds() to decide whether to sent transition events to magnifier. Originally it checks viewport scale > 1.0, but when scale = 1.0 the magnifier might be activatd and thus need to be notified when the events. Therefore, we change to check isForceShowingMagnifiableBounds(), which returns true iff the fullscreen magnifier is activated.

Bug: 271257917
Test: manually
Change-Id: I9c8fd68fa605a59ba0b2a4e082366ee27ed06aaa

Former-commit-id: 1e84af1aaa02f939728f6ad54e8d61ea56988e83