CSD: Fix setCsd TestApi logic
When executing the Gts interactive tests noticed that the csd was not
properly resetted. The main reason was because the timestamp was using
the global time instead of the boot time. This would cause that the next
entries would be deleted since they would be older than the CSD window.

Bug: 248566547
Test: atest GtsInteractiveAudioTestCases -- --module-arg
GtsInteractiveAudioTestCases: instrumentation-arg:ENABLE_MANUAL:=true
Change-Id: I002d4019768342ba3a2cca53b54a5b3d310b45c3

Former-commit-id: 645e1a0d904a8f9efeced85226cfdff76553ce76