Do not pass in anon callback.
The anon callback gets deferenced after the first few calls. Make sure
that it is a class field. Also update the documentation to reflect this.

There are no known bugs due to this, but proactively fixing this to
avoid unintended behavior.

Bug: 267821080
Test: add logs to see if the callbacks are called continuously.
onBiometricRunningStateChanged is called pretty much every time we go to
LS and leave LS.

Change-Id: If7fb2450f9025b85a30a10612966561c020ec865

Former-commit-id: e1bde69d4fba6989cdc4579885cf62c38af7206d