[Sb refactor] Use MobileMappings key for Unknown net types
ResolvedNetworkType.UnknownNetworkType was using "unknown" as the key,
but MobileMappings actually can potentially have some embedded
information about which network type icon to display in the case of
unknown. This would also always cause a lookup failure, causing the G
network indicator to show in cases where it should not be.

Test: MobileConnectionRepositoryTest
Bug: 269040632
Change-Id: I61f450faa756487983f300ff9dfefdcb74424e43

Former-commit-id: e5e9b8cfaff0cbfdcfc5b461841b30861eb1a5f6