Create an interface to update for view changes for bubbles
BubbleDataListener should be agnostic to the view that updates are being
sent to, this will make it easier to share the listener between
two different view implementations of bubbles (i.e. floating bubbles
and the bubble bar).

Also moves a method out of BubbleStackView which is better defined
in BubbleController.

Test: atest BubblesTest
Bug: 253318833
Change-Id: Ie6d85e25a50c3feeb21d468b6498f39680f354d4

Former-commit-id: e6de3e8fcca76d7ff1c896fcb5a0b92b6dc76368