Catch unusual DeadObjectExceptions.
This appears to be possible but very rare (0.0002% of sessions).
Launcher asks us to null out the LauncherUnlockAnimationController
in onDestroy(), but that's an IPC so might be delayed. Safest
to catch the exception here.

It should be fine to fail to play the unlock animation on a
Launcher that is being destroyed anyway.

Bug: 245867371
Test: atest SystemUITests
Change-Id: I96f14c2036daa945db324dc39b023a0facfb54dd

Former-commit-id: a72c7d90a856af311da2b0c71eab1506d2968dd9