Candidate Phase Metric Collector Data Type
Updated to match new atom, this handles the candidate phase. Note this
is *per provider* and thus we store this in a collection until the
logging occurs. How that happens may slightly change given our new
funnel points. In any case, this local collection of the metric will
help us emit properly.

Bug: 270403549
Test: To be chained in - are visually tested
Change-Id: Iad387c33f10c43b31308b18c339d72f9176a658e

Former-commit-id: 41893f958f193e1d3d0a6d525450ee7e71197698