Removed --verbose from `dumpsys activity`.
This option was added to provide extra but redundant information
in the ACTIVITY lines of dumpsys activity (it was redundant because
the information was also shown on the activity's TASK).

This CL removes that option and always display the info, as the
overhead is minimum and by doing so in a separate CL, it will be
easier to revert this behavior if neededed.

Test: adb shell dumpsys activity top| grep ACTIVITY | grep displayId || echo 'DOH'
Bug: 143499144

Change-Id: I58a58c5c2e751712595ff982b144697ca292767e

Former-commit-id: 4cb4ed0dc50535b2e49a3006f6ab5207f879e8c9