Placing InitialPhase Emit in Proper Locations
Note, some details, such as query start, can be optimized in future
versions. For example, we could use min(provider_start_1,
provider_start_2, ....) = query_start, and emit it. Alternatievely, we
could do this in the backend. For now, I've gone and made it a general
time that is un-emitted but exists as a utility locally. We may choose
to emit it in future versions as more granular backend queries are
developed.

Bug: 270403549
Test: To be chained in, builds for now
Change-Id: I6041bd6fd2d62c7ea501ccb3094a886dd9aa2758

Former-commit-id: 4f5d23740c76dcfcf855db8a2a9033bbb82fb700