Make the Factory Reset Protection unit tests work on TV
Some unit tests in LockscreenFrpTest and SyntheticPasswordTests are
broken on cf_x86_tv-userdebug because TV has
config_enableCredentialFactoryResetProtection set to false.  This isn't
a recent breakage -- it's been like this for a long time.

Fix this by mocking out the Resources so that the locksettings unit
tests work the same way on TV as on phone.

Bug: 272294822
Bug: 272295113
Test: atest com.android.server.locksettings # with cf_x86_tv-userdebug
Change-Id: I6296ca007112178867ed450057d5d2db8c94a09f

Former-commit-id: dd9add8ef345baf6582eb0e3fc830053d213e04c