Increase keyguardDrawnTimeout to reduce fold->AOD flickers
Once we hit the keyguard drawn timeout, the screen is turned on. However, in some cases sysui and launcher are not ready (due to big ui thread latency), and this results with the screen on while showing the wrong layout, followed by a rapid jump to the correct one.

This cl makes the timeout configurable in overlays for each device. In this way, the chances to have flicker are greatly reduced, but the underlying latency problem remains.
In parallel, we're looking to fix the root cause latency problem

Bug: 243015678
Test: trying to repro the scenario generating flicker many times, no flicker.
Change-Id: I08f147e7e4944656a0028df7e521441b1a9cc65c

Former-commit-id: 4a8ed3c57d917e2f82a29dc87f9f091297daa7e9