Prevent keyguard from dismissing if not...
dismissable and add a log here if this is the case.

KeyguardStateController#isLockscreenDismissable checks to see if the
user has unlocked with biometric auth or if the user's security method
is not secure (i.e. swipe/none). Is the user is attemping to dismiss the
lockscreen when the user has not bio auth unlocked and has a
pin/password/pattern, we also assert that the user has correctly auth'd
with the respective pin/password/pattern.

Fixes: 272249570
Test: Try to dismiss keyguard with bouncer and bio auth.
Test: sim pin/sim puk with swipe and pin security set.
Test: call StatusBarKeyguardViewManager#notifyKeyguardAuthenticated
directly when on LS and observe the log being reached.

Change-Id: I5150d58a44fcd927661672ec61b915f51801fb7f

Former-commit-id: 5def7eb4f9e7a2f79e251e9c876956dadbece5e6