ImeVisibilityApplier: fix hideIme no-op for Embedded display case.
Fix CL[1] uses the selfReportedDisplayId of the IME focused window
instead of the displayId to show IME that causes WMS#hideIme no-op
when the IME requester was on the embedded virtual display but the
actual IME control target was on the display.

Also, refine and remove previous TODO comments of invoking
WindowManagerInternal#hideIme part to make it clear about sending
displayId to WM is used to hide IME from the actual IME control
target of the target display.

[1]: I6e137a6e14f4fd7101322f54fda32ddcb5ccd017

Fix: 268266059
Test: atest DefaultImeVisibilityApplierTest#\
         testApplyImeVisibility_hideImeFromTargetOnSecondaryDisplay
Change-Id: Ie2380bcba44e44082f3cd044d4ec6c41c0193353

Former-commit-id: 65bc122ee9364a75392bf547061240caadd646d8