Removes comment dead link on MIME types
Removes a dead link on MIME types in Notification Messages, and
adds a pointer to suggest looking at ImageDecoder.isMimeTypeSupported
for information about what mime types are supported; notification
messages are decoded to drawables using ImageDecoder, so that list
should provide useful guidance. Other MIME types may be supported,
but the image types are likely to be most common, so ImageDecoder
seemed like a reasonable place to direct attention.

Change-Id: I474509d8874dc282ab8bcc00a9a6b86a6192d478
Test: Only changing comments
Bug: 154803285

Former-commit-id: f15072a854cfd1f77cd75b40cba3aa1fe4608077