Expand checkHealth() to look for more stalls.
The initial health check was only looking for processes stuck in the
"blocked" state for more than 10 minutes; this change expands that
to look for *any* broadcasts that have been waiting more than 10
minutes to be dispatched.

Note that the first broadcast matching this state will trigger the
wtf() event, and from that point onward we stop doing health checks
until the next reboot to avoid spamming dropbox.

Bug: 272685068
Test: treehugger
Change-Id: I6d84ccfecf72131195ed96498e2538e790e2bbd2

Former-commit-id: 05d59548e044834802858daf46603335409c6565