Checks SplitAttributes when reusing container
Before this CL, we checks whether a container can be reused based on
rules. However, there's a corner case that the SplitAttributesCalculator
function is updated before a new activity is started. It may lead to the
split pair has different splitAttributes even if activities matches the
same rule.

This CL also takes whether the calculated SplitAttributes match
the SplitAttributes of the container into account.

Bug: 241043377
Test: atest SplitAttributesCalculatorTest SplitControllerTest
Change-Id: I4bd4797d9462fccc9be56fd1cdb422e27a79bcff

Former-commit-id: 1af8ee5bd8b6d5df6bab3c23dcbfe7610ca2981b