Persist the launcher keep clear area.
This prevents the area from being lost due to other regions reporting
their respective keep clear areas. With shell transitions enabled
there is a possibility that the keep clear areas change between
onTaskAppearead and PipTransition#startAnimation callbacks, which would
cause the PiP to enter in the bottom right corner and then move above
Hotseat, instead of going directly to the correct position.

Test: e.g. atest ExpandPipOnDoubleClickTest
Test: the above test would fail otherwise, as the tapping happens at
the old position of PiP, not the one adjusted due to keep clear areas
Bug: 260779127
Bug: 271493738

Change-Id: I5537ce8c9b2395017eeb9e112f44994433733fe9

Former-commit-id: 96155ff2c682f8ba067946a7e3fd627c9de91634