Introducing AutomaticBrightnessStrategy
Creating an AutomaticBrightnessStrategy, which will be
responsible for managing all the auto-brightness related states when
evaluating the brightness.

This strategy is currently linked directly to DisplayPowerController2,
and is not implementing the DisplayBrightnessStrategy interface to keep
the scope limited. As a followup, this will be fully ported to inherit
all the properties of a strategy, and there will be no direct linkage
between DisplayPowerController and AutomaticBrightnessStrategy

Test: atest AutomaticBrightnessStrategyTest.java
Bug: 265415257
Change-Id: Idd8479109ee398729c8d84cab82a4bba5f0c6c1f

Former-commit-id: d6080c5a8743150e177d9ca2c85569ace34488d3