Eliminate delay before contacts lookup for a Notification
The RankingReconsideration produced by ValidateNotificationPeople had getDelay() == 1000 ms, which meant NMS waiting a full second before starting the work. This wasted half of the time available to decide whether a notification should produce sound during DND, for no good reason (despite what the comment claimed, this delay is not used as a timeout).

This CL also moves all the contacts-querying methods in ValidateNotificationPeople to PeopleRankingReconsideration. This makes it clear that the querying is only done as part of the reconsideration, and process() only queries the cache.

(This is a roll-forward with adjusted CTS tests).

Bug: 263839687
Test: atest
Change-Id: I80707fc1381410a5b64c3ac128e448d8d9a35b11

Former-commit-id: 8ee4b51d8d3dac448d277c18a8825195c940e5d3