Set app space position to app configuration
The frames in ClientWindowFrames for client use app coordinate space,
and InsetsState also apply app's scale. So the position of the bounds
in Configuration should be consistent with them.

So the sandbox calculation such as "displayFrame.left - bounds.left"
can be consistent. Previously it may be "appSpace - screenSpace".

In general, ActivityRecord#getBounds() is used for window layout in
server side so it is in screen space.
While record.getWindowConfiguration.getBounds() is used to report to
app so it is in app space.

Fix: 270401710
Test: SizeCompatTests#testFixedScreenBoundsWhenDisplaySizeChanged
      SizeCompatTests#testFixedOrientationRotateCutoutDisplay
Change-Id: Ife48bd5789db7cb9ba32302b345128cd4419a4b6

Former-commit-id: ccf766b921ee0fdfc64cb1cbf8896f5558cc741f