Update notification FSI logic for when device is considered dreaming
This change uses StatusBarStateController to check if the device is
dreaming, since that state is based on the dreaming broadcasts which are
only sent after the dream has finished waking up. This means that
StatusBarContoller.isDreaming will remain true during the wake-up phase
of the dream, therefore allowing full-screen intents in notifications.

This change also removes the RPC to DreamManagerService, since the dream
state is locally cached in StatusBarController

Fixes: 265977861
Test: atest StatusBarStateControllerImplTest
Test: atest NotificationInterruptStateProviderImplTest
Test: atest CentralSurfacesImplTest
Change-Id: I8e97fdb064c4b23517d69498638769aac26f419f
Merged-In: I8e97fdb064c4b23517d69498638769aac26f419f

Former-commit-id: 7a8aaaaad9307ed522585401023586bad63642c8