Fix DreamService -> DreamOverlayService synchronization issues.
This change moves the DreamOverlayService connection management logic
out of DreamService so that it can be separately tested. It also updates
the connection logic to use a handler for synchronization of messages,
and correctly cancels pending messages when needed.

This avoids consumers from running after the service has been
unbound.

Fixes: 267335683
Test: atest DreamOverlayConnectionHandlerTest
Test: flashed device and verified no crash when enabling/disabling
lowlight, and pressing the power button

Change-Id: I93855885ad80410693b8dcb4c2e696e0cabc5737

Former-commit-id: 4324e3418792df99e4e3181a24c24e2e0e930405