Reland "Use non-cached value when calculation button forced visible state"
This reverts commit ff6b98bc1f20c2fb73a47e8427ccff3ebf58eaed [formerly afe3acbbdec745f24b6ea31161ac4687d25c898d].

Previous landed patch caused nav bar added with a null insets frame
size, and there's a bug the insets size won't get updated at runtime if
it was null when added. Fixed the issue and reland the patch.

Necessary modification was made to make sure the givenContentInsets and
extended by cutout logic can be applied correctly.

Reason for revert: Reland with Fix of the flaky test

Test: SeamlessAppRotationTest
Test: Check setup wizard as stated in b/268210882
Test: DisplayPolicyTests
Test: DisplayPolicyInsetsTests
Bug: 268210882

Change-Id: I3511bd87852a696a2ec09fe71d1cde90b84a5b66

Former-commit-id: 6c2921def3c96d2b5d491d7624024c7188481c75