Run transition when app dies in freeform
In a freeform environment, currently a shell transition times out when
an app is killed via adb force-stop.

In this timeout case, setReady() isn't called where it's supposed to
because handleAppDied doesn't directly trigger any z-order or focus
update in freeform mode.

This CL adjusts the if condition in question so that this kind of
lingering transition will be executed even without any focus change.

This CL also fixes the timeout issue when dismissing placeholder activity and a PIP task is there.

Bug: 244123821
Bug: 268239637
fixes: 274536834
Test: adb shell am force-stop com.android.settings
Test: manual - Open settings + pip and fold the foldables

Merged-In: I07bdadd81f6992445b11e21dc08a68c231e8d5b8
Change-Id: I07bdadd81f6992445b11e21dc08a68c231e8d5b8

Former-commit-id: cdb03b309248bd36c36d54238c45958ddb1c4103