Add minimum height and width to recommendation covers
Found an illegal exception because the width or height of a certain
image view is zero. The crash came from adding the recommendation card.
So to make sure we have a certain value for height and width, we add
minWidth and minHeight attributes in style.And post set the image icon in order to avoid create bitmap before width and height are set.

Also, added a dimension for recommendation album icon size.

Bug: 271959514
Test: atest MediaControlPanelTest.
Test: checked recommendation card UI for old and new version.
Change-Id: Ic99cf0b5f58217c95940bae29fc3ba2e6efbbb9b

Former-commit-id: a08adafba9207eebb044dc605469853c85a11072