AS: only dispatch preferred mixer attributes change if the device is found.
When device is disconnected, only dispatch preferred mixer attributes
change if the device is found. When setting the device as connected, it
may not be cached by audio policy manager if there is any lower layer
issue happens while connecting the device. In that case, it may fail to
get AudioDeviceInfo for the disconnected device. To avoid crash,
AudioService should only notify preferred mixer attributes changed when
the disconnected device is found.

Bug: 273894204
Test: connect and disconnect USB devices
Change-Id: I4da7f7cc11336ea40f4f5dc0e042d781b1063733

Former-commit-id: fbaee6c0889702e5df4a9589d9689511cf088423