Add a delay to the removal of cached listener alarms
Sometimes a uid may quickly transition between process-states. To allow
for this without causing severe breakages, alarm maanger will wait for
a short duration before removing its exact listener alarms. If the uid
gets uncached during this time, the removal will be cancelled.

To start with, this delay is equal to the default freezer timeout but
this is left to be configurable via device_config.

Test: atest FrameworksMockingServicesTests:AppStateTrackerTest
Test: atest FrameworksMockingServicesTests:AlarmManagerServiceTest

Bug: 265195908
Change-Id: Ia1e57bf5163a0d543bd767ef5a44301331e7c48d

Former-commit-id: 8b6a9987f54fea3998d64334438ca558c8310eb5