Wait to restart Doze until after DreamMonitor starts.
Also, don't set the stored value to doze to false when we are starting
doze. Just leave it as true. It will get set to false if and when the
phone leaves the doze state.

Otherwise, if sysui dies in the moment between storing `false` and
actually starting doze, it can get left in a bad state.

Test: manually kill sysui a lot
Fixes: 268060399
Change-Id: I5cc60b30fbfc3c86da2f36b0f2789bb0c434339b

Former-commit-id: c969cfa0f7b2834ba2b4a90cedff9444c7e01b4d