Introduce TvPipMenuMode
This CL introduces a new IntDef - TvPipMenuMode. This allows to track
the current mode of the tv pip menu.

There are three states the the tv pip menu can be in:
1. No pip - in this mode only the pip tray is drawn (needed for rounded
   corners)
2. Move menu - in this mode the user can move the PiP to a different
   position on the screen
3. All actions mode - this mode presents the tv pip actions

Making the TvPipMenuController to have a single mode that the tv pip
menu can be in, makes it easier to reason about the state of the class
and easier to test.

Bug: 220108601
Test: atest TvPipMenuControllerTests
Change-Id: I95f3dfeaacacfc99e86125fa84554cd4af618417

Former-commit-id: 656d052e0d51561ee3a092d482a0bf0e5736819a