ActiveUnlock dismisses keyguard on trigger from assistant
The Assistant visibility doesn't change until the assistant
is attempting to address a query that requires auth. In this case,
without ActiveUnlock, the bouncer would show for the user to proceed,
and if the user enters their pin/patttern/password or biometric success,
then the user is brought to the secure assistant activity over the home
screen (keyguard is dismissed). To align with this CUJ, also make ActiveUnlock
dismiss the keyguard on successfull auth from a AU request coming from
the assistant visible signal.

Test: without biometrics enrolled, launch the assistant
with a query that requires auth to trigger active unlock.
When active unlock is triggered, the device will proceed to
enter the device (dismiss keyguard).
Test: atest KeyguardUpdateMonitorTest
Fixes: 269207635

Change-Id: Ic4f41b0705fb680691f7a3472853e76778cb9dfa

Former-commit-id: 10fb7116bdfa70528ba2fb1bebd45de3858bf949