[MMS] Don't show error for non sms app
Show "Switch to managed profile" dialog only in case of when message is
initiated from an SMS app in personal profile(and SIM user is associated
with managed profile).
We should not show this dialog in non SMS app case, as it may give
strange experience as seen in attached bug.

Bug: 268211439
Test: NA
Change-Id: Ie62ecca730312400b9ab759e8031433265b9f9cf

Former-commit-id: 395d193693aaebc5089c91bb375630b4d6e4d3c1