Revert "Adds assets, removes SFPS indicator, updates BP controller for rear display/tent mode"
This reverts commit 5e3c57bc89ad484d13b82b0de927b9906b4ee88a [formerly 566ca7306a8f3808817e61e1b2b222f539f2f496].

Reason for revert: This CL causes an issue where the SFPS indicator doesn't show up when expected, because it assumes the device to be in rear display mode. The fix for this ag/21801232 won't make it for the qpr-dev cut-off, so reverting this change and cherry picking it and the associated fix to udc.
Bug: 265198124
Change-Id: I5600640cbb0e3384ba46db6e9f26dc464ca444e0

Former-commit-id: 6363e0e3312c545f0671f1bf48e3834b0736f612