Match SFPS Keyguard color to the bouncer secondary color
The SFPS visual indicator should match the color of the delete
and enter buttons of the bouncer. This CL updates the colors to match.

Also send that the AUTH_REASON is from the keyguard when
the sideFps affordance is requested to show for the
alternate bouncer so that the themed colors get applied since
the biometric prompt & enrollment do NOT apply theme colors.

Test: manually bring up bouncer (both primary and alternate) and
see the SFPS affordance matches the color of the delete and enter
num key buttons on the PIN bouncer
Fixes: 270563217

Change-Id: I96d786cd2f035f0b936a3e1327b8a57652c2eb64

Former-commit-id: 38c21b2be2e106739438d515a3aa9b41148437c7