Fix the issue where resetting the refresh rates and brightness thresholds in DeviceConfig wouldn't make the system to fallback to the default values
Currently, if the DeviceConfig value is reset, we do not change the
various different types of brightness thresholds and refresh rates to the default values,
but rather gets stuck on the old stale value. The only was to recover from
this is the device reboot.

Bug: 265126371
Test: Manually changed the DeviceConfig values and confirmed the new
behaviour.
Test: atest DisplayModeDirectorTest

Merged-In: I700d8581975fe6ac0c0db4ec800f3698ffa0c2d1
Change-Id: I700d8581975fe6ac0c0db4ec800f3698ffa0c2d1

Former-commit-id: 3d75125a058ca1d4672266c843ed6435d4fb3bc2