Improve exit transition for task close case
We got task vanished callback after its transition animation
finished, and it lead to cause animation not smooth due to the
delay.

When task go to close, we will get a task info changed callback
first and the top activity is null in this callback task info.
We could use it to remove children task info early to reduce
the delay to dismiss split.

Fix: 263320869
Test: manual
Test: pass existing tests
Change-Id: I48486e46e539fe4647fc82a82b530d4637e3a544

Former-commit-id: 2a47e6a6ed4d5fdc812e9c1ce3f60b5ff226f31c