Ignore isInterruptive when sorting NotificationRecords
The existing behavior is:
* New notifications are always interruptive.
* Updated notifications are interruptive if and only if they change some subset of fields.
* We put interruptive notifications before non-interruptive ones when sorting (before considering their "when" time).

This leads to some unexpected behaviors, such as:
NM.notify(1, notif1);
NM.notify(2, notif2);
NM.notify(2, notif2);
placing notif1 before notif2, even though it's older and the third notify() call is a sorta-no-op.

This CL creates a flag to remove this specific comparation criterion so we can evaluate the behavior.

Bug: 204367726
Test: atest
Change-Id: I07404f0894d5a0d5592b25d70e67c4b2857f8bcf

Former-commit-id: 4630b64cfe7d59a80b79af0c04a98d4d71c1dc86