Use different pip shadow for each tv pip menu mode
This is the first of several CLs addressing the shadow on the tv pip
menu. In this CL we add a dedicated class TvPipBackgroundView for the
view hierarchy in the background surface of the Tv Pip menu. This
view is responsible for drawing the shadow behind the Tv PiP window.

The TvPipBackgroundView will be using a different shadow intensity for
each of the TvPipMenuModes. In this CL we adjust the shadow only using
the View translationZ attribute

Bug: 220108601
Test: atest TvPipMenuControllerTests
Change-Id: If191c4fac4a977f4678e59848aec2be6a2375e75

Former-commit-id: 4aa772aba141d2f18e863bb9568864735fd7d968