Fix UiAutomation for visible background users.
UiAutomation is "implemented" as an accessibility service, and the
A11Y infra doesn't support concurrent users (i.e., it only supports
the current user and its profiles). This CL works around that
limitation by temporarily setting the current user as the visible
background user while the UiAutomation connection is set.

Test: atest CtsUiAutomationTestCases android.app.cts.InstrumentationTest # CTS
Test: atest --user-type secondary_user_on_secondary_display CtsUiAutomationTestCases # CTS
Test: atest FrameworksCoreTests:android.app.InstrumentationTest FrameworksCoreTests:android.app.UiAutomationTest AccessibilityManagerServiceTest AccessibilitySecurityPolicyTest AccessibilityWindowManagerTest # unit

Fixes: 271188189
Change-Id: I8603e09e0bc8ee2aebb43dfe28dd3531cda05c37

Former-commit-id: ea241d52ed0bcf08253a3d9f8c137fa6ab77e08b