[SB Refactor] Define a `defaultConnections` flow in the shared repo.
This CL:
(1) Updates the wifi and mobile repos to use this shared flow instead
of defining their own default network callbacks.

(2) Updates `MobileConnectionsRepoImpl.defaultMobileNetworkConnectivity`
to have `isDefault = true` when mobile **or** carrier merged is the
default connection. This is needed to fix b/272586234: since carrier
merged is displayed as a mobile network, but we only display mobile
network information if mobile is default, we need `isDefault` to be true
even in the carrier merged case.

Bug: 272586234
Test: all tests in statusbar.pipeline
Test: manual: verify wifi icon still works and updates with normal wifi
Test: unfortunately, it's prohibitively difficult to set up a valid
carrier merged connection to test this, so I was not able to manually
test carrier merged. I wrote extensive unit tests instead

Change-Id: I7922b47547c3a4d186e78b76f7fae9cf3460d4d6

Former-commit-id: 0202e4a3373d81f5742d34cd925749800194421b