locksettings: allow {get,set}-disabled without providing current LSKF
There is no reason for the get-disabled and set-disabled commands to
require the current LSKF, so remove the requirement that it be provided.
This is safe for set-disabled because the lockscreen isn't actually
disabled until the LSKF is cleared, which must be done separately.

This avoids unlocking screen-lock-bound Keystore keys when not intended.

Bug: 273304787
Change-Id: I6472905c759a0e08e6bb59a59b51c5b40fe830ab

Former-commit-id: cb94dbee93d4a1407305c3484420af8eef39dad8