Don't move PiP if keyguard is showing or animating.
Without this there will be a PiP window shadow lingering around on the
screen, which looks like menu overlay, but it's actually the shadow.

Similar precaution is already implemented in the shelf height change
callback, so this is further adding the same logic to other places
that trigger movement, with or without keep clear areas flag.

Bug: 269402344
Test: manually, move PiP to the bottom, enter keyguard - no shadow

Change-Id: Id2b69c98c6a9d6afdc10d8f601c3e38a063084d0

Former-commit-id: 49b0ea6fecda66be07bed82a3c69f8545620b0f4