Add more data to removal history in alarm dump
Expanding the history to include more information about the alarm which
was removed. Specifically, the calculations for when the alarm is due is
extremely useful in debugging issues by looking at the dump.

Creating a mechanism to snapshot an alarm object for logging. More
fields can be added to this as needed.

Reversing the order of printing the removal history so it prints the
most recent removal first.

Test: Manually inspect the output of `dumpsys alarm`
Test: atest FrameworksMockingServicesTests:AlarmTest

Bug: 193063877
Change-Id: Id1baaa5c985a201ccd3d5f325c0781edec783fba

Former-commit-id: 61011ab3bea1c49acd4d61620ccdf3afe9583067