Implement flag to not relaunch on dock config change for apps without -desk resources
Activities are relaunched when they encounter an unhandled
configuration change. When a device is docked, this is surfaced to
activities as a uiMode change to UI_MODE_TYPE_DESK. Some apps don't
handle this configuration change by default, leading to a poor user
experience, such as video playback stopping when docking.

This change adds a flag that modifies this behavior and does not
relaunch activities if the device is docked or undocked and the app
does not have desk resources. Apps that have desk resources likely
already either handle the uiMode configuration update or expect to be
relaunched, so we want to preserve the default behavior for them. This
flag is turned off by default.

Bug: 266924897
Test: atest WmTests:ActivityRecordTests
Change-Id: Ib9eeb251f803e41ff8fd0ed56f1ea8d52e67ba81
Merged-In: Ib9eeb251f803e41ff8fd0ed56f1ea8d52e67ba81

Former-commit-id: 5fdd2f3ecfbb537d3c0421c0f1371a22796b1840