AS: Fix sometimes invalid volume set to LE Audio devices
In some cases when LE headet is connected, Audio Service can adjust
other devices volume level. We have to be sure that we use the
right volume level for the right device.

There seems to be a workaround for LE Audio devices, where the
volume level was not set for the outgoing calls. This does not seem
to be a problem anymore, and this workaround actually is the problem
and ocassionally can set volume level stored for the default speaker
device. This sometimes happens when rejecting the incoming call from
the LE Audio earbuds.

With this workaround removed, the issue with rejecting calls was not
reproduced anymore, and the issue with outgoing calls which this
workaround was suppose to prevent, was not observed as well.

It could be that the original issue that this woraround was made for
was already fixed.

Bug: 273165066
Test: tested manually making outgoing calls, taking and rejecting incoming calls
Change-Id: I97250c6cb2a7f3e59e51acbb9382bf03c26864ba

Former-commit-id: 6aa213d77ef6c1001fca0c5f889dfcc831bd804e