Give the wellbeing package a higher min balance.
The wellbeing app is pre-set on the device, not expected to be
interacted with much by the user, but can be expected to do work
in the background on behalf of the user. As such, it's a
pseudo-headless system app, so treat it as a headless system app
and give it a higher minimum satiated balance.

Bug: 267655220
Test: atest FrameworksMockingServicesTests:AlarmManagerEconomicPolicyTest
Test: atest FrameworksMockingServicesTests:CompleteEconomicPolicyTest
Test: atest FrameworksMockingServicesTests:JobSchedulerEconomicPolicyTest
Change-Id: I0c26a31e52c8b28b396781957b0fcecaef13020c

Former-commit-id: ca37e536aa700c6f3b1b3a49cb7c6818cea94770