Log manual rules in the zen rule snapshot dump.
This adds a parameter to explicitly indicate whether a rule to be dumped is actually a manual rule rather than using guesswork (that didn't work because manual rules don't actually have enabler set).

This change only catches when manual DND happens to be on at the time of the snapshot (roughly once per 12h), but it still provides more info than none.

Bug: 271469690
Test: atest ZenModeHelperTest, manual via statsd_testdrive
Change-Id: I76bc1d98990950de5515cdcedeff4be06dd3f4e8

Former-commit-id: 059f2203ec8de99db04519f1735605a48ef128b6