Allow package names to be retrieved for HotwordDetectionService
There are some system service we wish to provide to isolated_compute_app
to enable more efficient calculation in the isolated UID context.
HotwordDetectionService is currently one and the only one requiring this
capability. Some AGSA implemetions performs whitelisting on the package name thus we need to get the correct mapping of the uids.

Bug: 274535894
Test: sample app in isolated process can retrieve package name
Change-Id: Ieffa687c5bf6b2c8db5fbcf5761bc20676c7f1cf

Former-commit-id: 2066ebde190f8b7237037eedbc9bd5422b3a4d46