Write ViewCapture data to WmTrace when WindowManager dumps.
This allows apps like the Launcher to write ViewCapture data to the
wmtrace directory so it can be shown in their UI. Normally, Launcher
doesn't have the correct file permissions. Also, Launcher's dump method
is called after the wmtrace dir is written to the bug report, so we need
to dump sooner via a callback method invoked inside WindowManager.

Bug: 224595733
Test: Latency tested this change and verified that a bugreport generated
the file properly, moved it to the wmtrace directory, and was picked up
properly by the go/web-hv tool.

Change-Id: I9dc8e61070d2470354b79c1758103f9b2b00ac36

Former-commit-id: 3c492d8644ab4b4896244084abfe11ce6e47c6a6