Adds the gpu_counter_producer command.
This is a small binary that will load a GPU vendor provided shared
library called libgpudataproducer.so, look for the "start" symbol in
that library, and call it. This can be used to start the GPU hardware
counter Perfetto producer that will allow tracing of GPU profiling data
via Perfetto.

See go/gpu-profiling-arch for more details on the GPU profiling
architecture design.

Test: manual
Bug: b/273907878
Change-Id: Iacccb562d0ef154c105de034c8801642cab72532

Former-commit-id: 33da4aefce40bffca4f7bc39b9ee869e1430a7f0