Control behavior of unified restore pipeline after transport failure during K/V restore with flag
Currently unified restore pipeline finishes immediately if restore of a package fails due to transport error. This is mostly fine as transport level errors usually happen when transport is broken.
But there are some situations where such an error is package specific - like when system app goes from KV backups to full backups, and the bundled version of the app is unable to handle Full data restore (as bundled version could be older version).

To take this into account, we allow the restore pipeline to continue, when a transport level error happens in K/V restore, based on the value of flag - `unified_restore_should_continue_after_transport_level_failure_in_keyvalue_restore`

Bug: 128499560
Test: atest BackupAndRestoreFeatureFlagsTest
	atest PerformUnifiedRestoreTaskTest
Change-Id: I799d259387bd914877066920b8168fd11b5bfd45

Former-commit-id: cedd0bec7dceaf038d26fff91065f6b827dff11a