Give headless system apps a higher min balance.
Pre-installed apps without a "front door" activity (an activity that
is directly from the launcher) won't be interacted with much by the
user, but can be expected to do work in the background on behalf of
the user. As such, give these apps a higher minimum satiated balance
so they can do the work they need to.

Bug: 202954395
Test: atest FrameworksMockingServicesTests:AlarmManagerEconomicPolicyTest
Test: atest FrameworksMockingServicesTests:CompleteEconomicPolicyTest
Test: atest FrameworksMockingServicesTests:JobSchedulerEconomicPolicyTest
Change-Id: I79b85c808adcba017de092525d5a0084f7c10834

Former-commit-id: f1e9e1b25a2e6f872ba339a024d84fa168f18fe9