Always report keyguard-going-away even if empty
Normally we abort all empty (no-op) transitions. It doesn't actually
make sense to have a keyguard-going-away without something else to
show under it, but it seems like poorly-timed app logic can result
in a keyguard-going-away transition with nothing showing.

So change the logic to always report keyguard-going-away regardless
of whether it is empty or not.

Bug: 274031768
Test: ShellTransitionTests#testEmptyTransitionStillReportsKeyguardGoingAway
Change-Id: I7f92d198b691c5bfcf1cc9d84e7b2a8913523c18

Former-commit-id: e6b3aa3dbf1027e8e8c821b9caa5a21004924977