Fix text clipping of EmphasizedNotificationButton, when the system font size is at maximum
Fixes the visually clipped text of EmphasizedNotificationButton (which is used in sticky HUNs, CallStyleNotifications, etc.), when the Accessibility - Display size and text - font size is at its maximum(2.0 scale).

Test: Manual.
Go to Settings > Accessibility > Display size and text. Adjust the "Font size" to the biggest.
Post a sticky HUN, and a CallStyleNotification.
The action button texts shouldn't be clipped either when the notifications are showing as HUNs, or in the Shade.
Bug: 273401782
Bug: 274783980

Change-Id: Id87bb579fe966de832b0f95a2a950a2778217031

Former-commit-id: 8be486f83b277842ef5face04b75c127b53d2ab9