Remove pending runnables in QSTileView#changeState
When the view is not attached to a window, the runnables accumulate in
the view mRunQueue. If the view is never attached (for example, view in
the third page of QS), many of these can accumulate.

To fix, create a special runnable such that all instances are equal to
each other (so we can remove them). We cannot use the handler methods to
tag the runnable with a token because the problem arises when the view
is not attached (and therefore, there's no handler).

Test: manual, tiles get updated
Test: manual, using heapdump, check that runnables don't accumulate
Fixes: 269252339
Change-Id: I5371a19e1daf7b68eb3b152c5aa964774a372ebd

Former-commit-id: 6a22e3710fe0751048f753f42a90074ede6b08ec