Fix translucent in smart linking
This CL fixes the problem with letterboxing for
activities already in history relaunched from a
smart linking using a translucent trampoline activity.
The problem is related to a wrong behaviour with
double-tap which just moves some of the activities
in the task and not all of them.
Ignoring not visible activities as candidate to be
the first opaque activity beneath, fixes the problem.

Fixes: 274765111
Test: Run `atest SizeCompatTests`

Change-Id: I9e0f4f4ff64c6666903a3df5ca162aa5c2705ce6

Former-commit-id: 5aea1095a2eee9e5b0cc767e632cba0679d9556e