Request transition for corner case of keyguard state
If an app requests to occlude keyguard after its first resumed and
relayout is done (outside scope of UnknownAppVisibilityController),
the transition may not collect any visibility change.

Then when setShowWhenLocked takes effect to notify keyguard, there
won't be transition when calling setKeyguardShown. And the screen may
be either blank because the launching activity's parent was hidden
or keyguard is still showing because there is no transition to tell
keyguard to hide by TRANSIT_FLAG_KEYGUARD_LOCKED.

Also
- Reduce noise log when making device sleep
- Recover the surface visibility of parents if the activity
  is changing to visible outside a transition.

Bug: 274123143
Test: Turn off display or on lockscreen.
      Cold/warm/hot launch an activity which calls
      setShowWhenLocked and setTurnScreenOn with a delay.
      The visibility error log should not appear and
      the activity is visible without lockscreen.

Change-Id: I1fe16af079564a99cd61b0e3accf8699bc72e1e6

Former-commit-id: 8eadecca2206d7bd90dbd6404d4e139828b05fe8