[multi-shade] Fix multi touch bug.
Dragging down with one finger, adding a second finger, then removing the
original finger was crashing because the pointer index for the saved off
pointer ID is missing.

This CL fixes it by seamlessly moving over to the next pointer when the
original pointer is lifted.

Bug: 274159734
Test: manually verified that (a) the crash doesn't happen and (b) the
shaded continues to drag down smoothly without a "jump".
Test: I considered a unit test but decided against it because it was going to be very hard to set up MotionEvent instances for different pointers based on the obtain API.

Change-Id: Idbc534660b947ce450aa819e1322eb934978de79

Former-commit-id: 17ced90e60a667475b2236ada32f22d4ce4a8ab7