Fix dismiss callback when keyguard is dismissing.
AnimatingAway is set to false before hide is called, causing the dismiss
action to be cancelled. This is breaking CarSysUI. Maintain parity with
the legacy bouncer by setting disppear animation to null when we hide
the bouncer so we maintain the state of animating away when we check to
see if we should dismiss the callback.

Fixes: 271166101
Test: Atest
Change-Id: Ia51ec982dd78b69db82f69ae525e402bd461f294
Merged-In: Ia51ec982dd78b69db82f69ae525e402bd461f294

Former-commit-id: 74f6dce9851cafdccd4281dfd0086bb577a7c7b6