fix(non linear font scaling): add a setLineHeight() that accepts explicit units.
This is so the TextView can properly calculate a proportional line
height when non-linear font scaling is in effect. Otherwise, a desired
2x line height at font scale 1.0 will not be 2x at font scale 2.0, due
to non-linear font scaling compressing higher SP sizes.

It also better mirrors the setTextSize() API.

The actual fix for calculating the proportional sizes will come in a
subsequent CL.

Bug: 273326061
Test: atest cts/tests/tests/widget/src/android/widget/cts/TextViewTest.java

Change-Id: I79d74ff4f5fe1886ca2ed9c4df708be484fd888f

Former-commit-id: ba9126c49bfb4eba0bad1cd4dbae5fc24a3fd1d8