[multi-shade] Remove shade content padding when the shade is collapsed.
Having non-zero padding around the shade content when the shade is
collapsed actually protrudes into the display and swallows clicks on the
user switcher chip in the status bar near the top of the display.

This CL fixes that by calculating the padding dynamically such that it's
never bigger than 12dp or the height of shade, whichever is smaller.

Bug: 274159734
Test: manually verified that I can click on the user switcher chip in
the status bar when the shade is collapsed. I even used a stylus for
increased accuracy. I also verified that the shade can be expanded and
collapsed normally both from outside of it and inside of it.

Change-Id: I5fab6d9ab6f0e1d771866fcb0330a6c44846d89b

Former-commit-id: c270be975cf1a2d5bfd34db1c279715533330ab4